pyproject.toml
[tool.poetry]
name = "koffee"
version = "0.1.0"
description = "A Korean<>English video file translator."
authors = ["Kim, Andrew <andrewkimka@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
koffee = "koffee.cli:main"

[tool.poetry.dependencies]
python = ">=3.10"
faster-whisper = "^1.0.3"
transformers = "^4.39.3"
ffmpeg-python = "^0.2.0"
torch = "^2.2.2"
accelerate = "^0.29.3"
cyclopts = "^2.9.9"
rich = "^13.9.2"
pydantic = "^2.9.2"
sentencepiece = "^0.2.0"
sacremoses = "^0.1.1"

[tool.poetry.group.dev.dependencies]
coverage = "^7.9.1"
mypy = "^1.16.1"
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
pytest-mock = "^3.14.1"
ruff = "^0.12.0"
behave = "^1.2.6"

[tool.coverage.run]
branch = true
source = ["src"]

[tool.coverage.report]
show_missing = true
fail_under = 80

[tool.pytest.ini_options]
testpaths=["tests"]

[tool.pyright]
include = ["src"]

[tool.mypy]
files = ["src", "tests"]
cache_dir = "${HOME}/.mypy_cache"
ignore_errors = false
disallow_untyped_defs = true

[tool.ruff]
target-version = "py311"
line-length = 88
fix = true
unsafe-fixes = true

lint.select = [
  "E",   # pycodestyle errors
  "F",   # pyflakes
  "I",   # isort
  "UP",  # pyupgrade
  "B",   # bugbear
  "SIM", # simplify
  "D",   # pydocstyle
  "PL",  # pylint-like
  "C90", # mccabe
]

lint.ignore = []

exclude = ["tests/data", ".venv", "build", "dist"]

[tool.ruff.lint.pydocstyle]
convention = "google"  # or "numpy" or "pep257"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
